Rewrote InstallFileFromFomod in ModInstallScript chasing issue 7.  Details @ http://git.loverslab.com/prideslayer/fomm/issues/7
Changes to Archive.cs and Fomod.cs were required to make some properties and methods public rather than private or protected, since the bulk of the rewrite is a copy/paste job from the functions IFFF was calling.

Ver bump to 10.  This commit represents v10b3, will tag as such.